requirements.txt
flask
flask-wtf
flask-login
flask-bcrypt
flask-mail
flask-sqlalchemy
pillow
itsdangerous
gunicorn
python-dotenv
